chore(build): retry flakey data-table e2e test
