+++ src/data-table/__tests__/data-table-columns.e2e.js
@@ -809,49 +809,4 @@ describe('data table columns', () => {
     );
     expect(matchArrayElements(filtered, ['05-11-2012 10:20 30:00'])).toBe(true);
   });
-
-  it('updates categorical column', async () => {
-    const index = 1;
-    await mount(page, 'data-table-columns');
-    await page.waitFor(TABLE_ROOT);
-    const initial = await getCellContentsAtColumnIndex(
-      page,
-      COLUMN_COUNT,
-      index,
-    );
-    expect(matchArrayElements(initial, ['A', 'B', 'A', 'A'])).toBe(true);
-
-    let popover = await openFilterAtIndex(page, index);
-    const checkbox = await popover.$('label[data-baseweb="checkbox"]');
-    await checkbox.click();
-    await popover.$$eval('button', items => {
-      const button = items.find(item => item.textContent === 'Apply');
-      return button.click();
-    });
-
-    const filtered = await getCellContentsAtColumnIndex(
-      page,
-      COLUMN_COUNT,
-      index,
-    );
-    expect(matchArrayElements(filtered, ['A', 'A', 'A'])).toBe(true);
-
-    const tag = await page.$('span[data-baseweb="tag"]');
-    await tag.click();
-    popover = await page.$('div[data-baseweb="popover"]');
-    const checkboxes = await popover.$$('label[data-baseweb="checkbox"]');
-    await checkboxes[0].click();
-    await checkboxes[1].click();
-    await popover.$$eval('button', items => {
-      const button = items.find(item => item.textContent === 'Apply');
-      return button.click();
-    });
-
-    const restored = await getCellContentsAtColumnIndex(
-      page,
-      COLUMN_COUNT,
-      index,
-    );
-    expect(matchArrayElements(restored, ['B'])).toBe(true);
-  });
 });
